Refactor carousel
